+++ src/hooks/index.ts
@@ -4,3 +4,4 @@ export { useLoadAquariumDetails } from './useLoadAquariumDetails';
 export { useLoadAquariums } from './useLoadAquariums';
 export { useResetParams } from './useResetParams';
 export { useSelectedAquarium } from './useSelectedAquarium';
+export { useSignUp } from './useSignUp';
